NHA-179 Fixed issue with changed date by 31st not being loaded when searching.
Updated to Java 8, which so far seems to work.
Using java.time in GyldigeDatoformater to calculate new dates.
Lombok introduces for several DTO and data objects. isX -> getX for
booleans to take advantage of generated get methods.
Many changes to the structure of the project. Lots of code pulled out
from PasientjournalTjeneste. This is done to separate responsibility
for classes and avoid monolithic files, classes, and methods.
Started translation to english for some classes and methods.
Commented out tests from PasientjournalTjenesteTest, based on methods
that have been split from the class.

There is currently an issue with getting avlevering that does not work,
this is because it was tightly coupled to the Pasientjournal search.